move flow_controller to pkg/resource.
